third_party/WebKit/LayoutTests/typedcssom/cssLengthValue.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script>

test(function() {
  var result = CSSLengthValue.from(5, 'px');
  assert_equals(result.constructor.name, CSSSimpleLength.name);
  assert_equals(result.value, 5);
  assert_equals(result.type, 'px');
}, "Test that CSSLengthValue's static from(double value, LengthType type) method produces a CSSSimpleLength.");

test(function() {
  var result = CSSLengthValue.from({px: 1, percent: 2.5});
  assert_equals(result.constructor.name, CSSCalcLength.name);
  assert_equals(result.px, 1);
  assert_equals(result.percent, 2.5);
}, "Test that CSSLengthValue's static from(CalcDictionary dictionary) method produces a CSSCalcLength.");

</script>
<body>
</body>
